Refactor: Extract new async fn `ResolutionInstallFromSource::install`
